Refactor UI utility methods
... into UIUtils, and move UIUtils into the "util" package
