+++ GPSTest/src/main/java/com/android/gpstest/NavigationDrawerFragment.java
@@ -41,6 +41,7 @@ import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import com.android.gpstest.util.UIUtils;
 import com.android.gpstest.view.ScrimInsetsScrollView;
 
 import java.util.ArrayList;
