--- GPSTest/src/main/java/com/android/gpstest/UIUtils.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2018 Sean J. Barbeau
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.android.gpstest;
-
-import android.os.Build;
-import android.view.View;
-
-/**
- * Utilities for processing user inteface elements
- */
-
-public class UIUtils {
-
-    /**
-     * Formats a view so it is ignored for accessible access
-     */
-    public static void setAccessibilityIgnore(View view) {
-        view.setClickable(false);
-        view.setFocusable(false);
-        view.setContentDescription("");
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            view.setImportantForAccessibility(View.IMPORTANT_FOR_ACCESSIBILITY_NO);
-        }
-    }
-}
