+++ GPSTest/src/main/java/com/android/gpstest/util/UIUtils.java
@@ -1,0 +1,101 @@
+/*
+ * Copyright (C) 2018 Sean J. Barbeau
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.android.gpstest.util;
+
+import android.app.Activity;
+import android.content.Context;
+import android.os.Build;
+import android.support.v4.app.Fragment;
+import android.view.View;
+
+import java.util.concurrent.TimeUnit;
+
+/**
+ * Utilities for processing user inteface elements
+ */
+
+public class UIUtils {
+
+    /**
+     * Formats a view so it is ignored for accessible access
+     */
+    public static void setAccessibilityIgnore(View view) {
+        view.setClickable(false);
+        view.setFocusable(false);
+        view.setContentDescription("");
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+            view.setImportantForAccessibility(View.IMPORTANT_FOR_ACCESSIBILITY_NO);
+        }
+    }
+
+    /**
+     * Converts screen dimension units from dp to pixels, based on algorithm defined in
+     * http://developer.android.com/guide/practices/screens_support.html#dips-pels
+     *
+     * @param dp value in dp
+     * @return value in pixels
+     */
+    public static int dpToPixels(Context context, float dp) {
+        // Get the screen's density scale
+        final float scale = context.getResources().getDisplayMetrics().density;
+        // Convert the dps to pixels, based on density scale
+        return (int) (dp * scale + 0.5f);
+    }
+
+    /**
+     * Returns true if the activity is still active and dialogs can be managed (i.e., displayed
+     * or dismissed), or false if it is not
+     *
+     * @param activity Activity to check for displaying/dismissing a dialog
+     * @return true if the activity is still active and dialogs can be managed, or false if it is
+     * not
+     */
+    public static boolean canManageDialog(Activity activity) {
+        if (activity == null) {
+            return false;
+        }
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
+            return !activity.isFinishing() && !activity.isDestroyed();
+        } else {
+            return !activity.isFinishing();
+        }
+    }
+
+    /**
+     * Returns true if the fragment is attached to the activity, or false if it is not attached
+     *
+     * @param f fragment to be tested
+     * @return true if the fragment is attached to the activity, or false if it is not attached
+     */
+    public static boolean isFragmentAttached(Fragment f) {
+        return f.getActivity() != null && f.isAdded();
+    }
+
+    /**
+     * Returns a human-readable description of the time-to-first-fix, such as "38 sec"
+     *
+     * @param ttff time-to-first fix, in milliseconds
+     * @return a human-readable description of the time-to-first-fix, such as "38 sec"
+     */
+    public static String getTtffString(int ttff) {
+        if (ttff == 0) {
+            return "";
+        } else {
+            return TimeUnit.MILLISECONDS.toSeconds(ttff) + " sec";
+        }
+    }
+}
